주문 추가

+++ src/main/java/com/example/mykioskbackend/menucategory/MenuCategoryRepository.java
@@ -2,5 +2,5 @@ package com.example.mykioskbackend.menucategory;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuCategoryRepository extends JpaRepository<MenuCategory, Integer> {
+public interface MenuCategoryRepository extends JpaRepository<MenuCategory, Long> {
 }
